Allow float weight in NewSave schema
Change weight type from int to float in the NewSave Pydantic model to accept decimal weights while preserving validation (ge=0, le=1000). This makes the diary schema support fractional weights (e.g., 12.5). Note: clients that assumed integer weights may need updates to handle floats.

+++ backend/src/schemas/diary.py
@@ -10,7 +10,7 @@ ExerciseName = Annotated[str, StringConstraints(strip_whitespace=True, min_lengt
 class NewSave(BaseModel):
     exer_id: FirestoreId
     exerName: ExerciseName
-    weight: int = Field(ge=0, le=1000)
+    weight: float = Field(ge=0, le=1000)
     reps: int = Field(ge=1, le=1000)
